[FIX] File Migration

+++ database/migrations/2026_01_25_144601_create_nilai_ekonomi_details_table.php
@@ -10,15 +10,17 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::create('nilai_ekonomi_details', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('nilai_ekonomi_id')->constrained('nilai_ekonomi')->onDelete('cascade');
-            $table->foreignId('commodity_id')->constrained('m_commodities');
-            $table->decimal('production_volume', 15, 2);
-            $table->string('satuan')->default('Kg');
-            $table->decimal('transaction_value', 19, 2);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('nilai_ekonomi_details')) {
+            Schema::create('nilai_ekonomi_details', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('nilai_ekonomi_id')->constrained('nilai_ekonomi')->onDelete('cascade');
+                $table->foreignId('commodity_id')->constrained('m_commodities');
+                $table->decimal('production_volume', 15, 2);
+                $table->string('satuan')->default('Kg');
+                $table->decimal('transaction_value', 19, 2);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
